removida controlPreference
