+++ AnkiDroid/src/main/res/layout/control_preference.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context="com.ichi2.preferences.ControlPreference2"
+    tools:context="com.ichi2.preferences.ControlPreference"
     >
 
     <com.ichi2.ui.FixedTextView
