+++ AnkiDroid/src/main/res/values/10-preferences.xml
@@ -252,7 +252,6 @@
     <string name="binding_add_gesture">Add gesture</string>
     <string name="binding_replace_gesture">Replace gesture</string>
     <string name="binding_add_key">Add key</string>
-    <string name="binding_replace_key">Replace key</string>
     <string name="key_picker_default_press_key">Press a key</string>
     <string name="binding_add_axis">Add joystick/motion controller</string>
     <string name="axis_picker_move_joystick">Move a joystick/motion controller</string>
